Fix compiler warnings for C23 compatibility
- Fix braced-scalar-init warnings: replace {{0}} with {0} for array
  sentinels in SlotInfo and MethodInfo arrays
- Fix function prototype: add proper parameters to meth_field_cycleColors()
  declaration in cl_field.h
- Fix unused-value warnings in bright.c: use *(destptr++) instead of
  separate assignment and increment
- Fix function prototype: add parameters to g3_bitson() forward declaration
- Fix type conversions: change int to size_t for strlen() results in
  path.c, misc.c, mystrings.c
- Fix type conversions: add explicit casts for XTextItem.nchars (int)
  and pointer differences (ptrdiff_t)
- Fix sign-conversion warning: add explicit cast in strToVal() loop

+++ src/viola/glib_x.c
@@ -1407,7 +1407,7 @@ int GLPaintText(Window w, GC gc, int fontID, int x0, int y0, char* str)
 {
     XTextItem item;
     item.chars = str;
-    item.nchars = strlen(str);
+    item.nchars = (int)strlen(str);
     item.delta = 0;
     item.font = FontFont(fontID);
